requirements.txt
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-bigquery
google-cloud-core
google-crc32c
google-resumable-media
googleapis-common-protos